+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/Appointment/BoolFilters/Upcoming.php
@@ -5,15 +5,14 @@ declare(strict_types=1);
 namespace Espo\Modules\EspoDental\Classes\Select\Appointment\BoolFilters;
 
 use DateTimeImmutable;
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\Modules\EspoDental\Entities\Appointment;
 use Espo\ORM\Query\Part\WhereClause;
 use Espo\ORM\Query\Part\WhereItem;
 
-class Upcoming implements Filter
+class Upcoming extends RawBoolFilter
 {
-    public function apply(User $user): ?WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return WhereClause::fromRaw([
             'dateStart>=' => (new DateTimeImmutable())->format('Y-m-d H:i:s'),