+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/AssistantActionProposal/BoolFilters/PendingReview.php
@@ -4,15 +4,14 @@ declare(strict_types=1);
 
 namespace Espo\Modules\EspoDental\Classes\Select\AssistantActionProposal\BoolFilters;
 
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\Modules\EspoDental\Entities\AssistantActionProposal;
 use Espo\ORM\Query\Part\WhereClause;
 use Espo\ORM\Query\Part\WhereItem;
 
-class PendingReview implements Filter
+class PendingReview extends RawBoolFilter
 {
-    public function apply(User $user): WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return WhereClause::fromRaw(['status' => AssistantActionProposal::STATUS_PENDING_REVIEW]);
     }