+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/Common/UserAwareRawBoolFilter.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Espo\Modules\EspoDental\Classes\Select\Common;
+
+use Espo\Entities\User;
+
+abstract class UserAwareRawBoolFilter extends RawBoolFilter
+{
+    public function __construct(protected User $user)
+    {
+    }
+}