+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/Invoice/BoolFilters/OnlyOverdue.php
@@ -5,15 +5,14 @@ declare(strict_types=1);
 namespace Espo\Modules\EspoDental\Classes\Select\Invoice\BoolFilters;
 
 use DateTimeImmutable;
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\Modules\EspoDental\Entities\Invoice;
 use Espo\ORM\Query\Part\WhereClause;
 use Espo\ORM\Query\Part\WhereItem;
 
-class OnlyOverdue implements Filter
+class OnlyOverdue extends RawBoolFilter
 {
-    public function apply(User $user): ?WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return WhereClause::fromRaw([
             'status' => [Invoice::STATUS_ISSUED, Invoice::STATUS_PARTIAL_PAID],