+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/Material/BoolFilters/LowStock.php
@@ -4,15 +4,14 @@ declare(strict_types=1);
 
 namespace Espo\Modules\EspoDental\Classes\Select\Material\BoolFilters;
 
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\Modules\EspoDental\Entities\Material;
 use Espo\ORM\Query\Part\WhereClause;
 use Espo\ORM\Query\Part\WhereItem;
 
-class LowStock implements Filter
+class LowStock extends RawBoolFilter
 {
-    public function apply(User $user): ?WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return WhereClause::fromRaw([
             'stockLevel' => [Material::LEVEL_LOW, Material::LEVEL_CRITICAL, Material::LEVEL_OUT],