+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/NotificationLog/BoolFilters/OnlyFailed.php
@@ -4,15 +4,14 @@ declare(strict_types=1);
 
 namespace Espo\Modules\EspoDental\Classes\Select\NotificationLog\BoolFilters;
 
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\Modules\EspoDental\Entities\NotificationLog;
 use Espo\ORM\Query\Part\WhereClause;
 use Espo\ORM\Query\Part\WhereItem;
 
-class OnlyFailed implements Filter
+class OnlyFailed extends RawBoolFilter
 {
-    public function apply(User $user): ?WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return WhereClause::fromRaw(['status' => NotificationLog::STATUS_FAILED]);
     }