+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/SalaryEntry/BoolFilters/ThisMonth.php
@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Espo\Modules\EspoDental\Classes\Select\SalaryEntry\BoolFilters;
 
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\Modules\EspoDental\Classes\Select\Common\DateRangeHelper;
 use Espo\ORM\Query\Part\WhereItem;
 
-class ThisMonth implements Filter
+class ThisMonth extends RawBoolFilter
 {
-    public function apply(User $user): ?WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return DateRangeHelper::thisMonth('periodFrom');
     }