+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/SalaryProfile/BoolFilters/OnlyActive.php
@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Espo\Modules\EspoDental\Classes\Select\SalaryProfile\BoolFilters;
 
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\ORM\Query\Part\WhereClause;
 use Espo\ORM\Query\Part\WhereItem;
 
-class OnlyActive implements Filter
+class OnlyActive extends RawBoolFilter
 {
-    public function apply(User $user): ?WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return WhereClause::fromRaw(['isActive' => true]);
     }