+++ src/files/custom/Espo/Modules/EspoDental/Classes/Select/Visit/BoolFilters/Today.php
@@ -4,14 +4,13 @@ declare(strict_types=1);
 
 namespace Espo\Modules\EspoDental\Classes\Select\Visit\BoolFilters;
 
-use Espo\Core\Select\Bool\Filter;
-use Espo\Entities\User;
+use Espo\Modules\EspoDental\Classes\Select\Common\RawBoolFilter;
 use Espo\Modules\EspoDental\Classes\Select\Common\DateRangeHelper;
 use Espo\ORM\Query\Part\WhereItem;
 
-class Today implements Filter
+class Today extends RawBoolFilter
 {
-    public function apply(User $user): ?WhereItem
+    protected function buildWhereItem(): WhereItem
     {
         return DateRangeHelper::today('startedAt');
     }